requirements.txt
django==2.2.7
requests
djangorestframework
pymongo
django-cors-headers